requirements.txt
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
dnspython==1.16.0
google-ads==3.3.0
google-api-core==1.16.0
google-api-python-client==1.7.11
google-auth==1.11.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.26.0
httplib2==0.17.0
idna==2.8
mysql-connector-python==8.0.19
numpy==1.18.1
oauth2client==4.1.3
oauthlib==3.1.0
pandas==0.24.2
protobuf==3.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygsheets==2.0.2
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
uritemplate==3.0.1
urllib3==1.25.8
